# url-shortener/backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: URL Shortener API
  description: A simple URL shortening service.
  version: 0.1.0
servers:
  - url: /api/v1
paths:
  /links:
    post:
      operationId: createLink
      summary: Create a short link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLinkRequest"
      responses:
        "201":
          description: Link created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Link"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Custom alias already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      operationId: listLinks
      summary: List short links
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: A page of links
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LinkList"
  /links/{code}:
    parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getLink
      summary: Get a short link by code
      responses:
        "200":
          description: The link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Link"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteLink
      summary: Delete a short link by code
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /links/{code}/events:
    parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: listLinkEvents
      summary: List access events for a short link
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: A page of access events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventList"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
    CreateLinkRequest:
      type: object
      required:
        - url
      properties:
        url:
          type: string
          format: uri
          description: The long URL to shorten.
          example: https://example.com/some/very/long/path
        custom_alias:
          type: string
          description: Optional custom short code. If omitted, one is generated.
          pattern: "^[A-Za-z0-9_-]{3,32}$"
          example: my-link
    Link:
      type: object
      required:
        - code
        - url
        - short_url
        - visit_count
        - created_at
      properties:
        code:
          type: string
          example: aB3xY9
        url:
          type: string
          format: uri
          example: https://example.com/some/very/long/path
        short_url:
          type: string
          format: uri
          example: http://localhost:8080/aB3xY9
        visit_count:
          type: integer
          format: int64
          example: 42
        created_at:
          type: string
          format: date-time
    LinkList:
      type: object
      required:
        - items
        - total
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Link"
        total:
          type: integer
          format: int64
    Event:
      type: object
      required:
        - accessed_at
        - referer
        - user_agent
      properties:
        accessed_at:
          type: string
          format: date-time
        referer:
          type: string
          example: https://news.example/article
        user_agent:
          type: string
          example: Mozilla/5.0
    EventList:
      type: object
      required:
        - items
        - total
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        total:
          type: integer
          format: int64
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
        code:
          type: string