# url-shortener/backend/internal/repository/dbgen/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package dbgen

type Link struct {
	Code      string
	Url       string
	CreatedAt string
}

type LinkEvent struct {
	ID         int64
	Code       string
	AccessedAt string
	Referer    string
	UserAgent  string
}